Revert "Many changes because database was created and a rebuild for java class was necessary"
This reverts commit 4d3832fc0965d2f16bf8288753d88ab4855f5471.

+++ PetCitySoft/src/main/java/co/com/carp/petcity/entity/PetVisitInfo.java
@@ -17,7 +17,7 @@ public class PetVisitInfo {
 	/**
 	 * Pet's weight
 	 */
-	private float weight;
+	private double weight;
 	
 	/**
 	 * Pet's food
@@ -46,14 +46,14 @@ public class PetVisitInfo {
 	/**
 	 * @return the weight
 	 */
-	public float getWeight() {
+	public double getWeight() {
 		return weight;
 	}
 
 	/**
 	 * @param weight the weight to set
 	 */
-	public void setWeight(float weight) {
+	public void setWeight(double weight) {
 		this.weight = weight;
 	}
